refactor: move pairDisplayRow to the leaf module; dedup its consumers
/simplify pass (4 cleanup agents: reuse / simplification / efficiency / altitude).

- altitude: pairDisplayRow was a PURE (info, meta) function parked in
  useTerminalStore.ts (a composition root), one layer above two of its natural
  consumers — buildWorkspaceEntries (dockModel) and CanvasMinimap can't import
  useTerminalStore without a dependency inversion, so they re-spelled the gate
  inline. Move it (with no behavior change) into terminalDisplay.ts, the pure
  leaf module every consumer already imports, and re-point TerminalMeta / RowPips.
- reuse + simplification: MobilePullChrome hand-composed two near-identical
  accessors (activeInfo / activeMeta) that each re-read activeId(). Collapse to
  one paired `activeRow` accessor — reads activeId() once, gates both-present
  once via the shared pairDisplayRow (same coherent pair the desktop header gets).
- buildWorkspaceEntries now gates through pairDisplayRow instead of an inline
  `!info || !meta`.

Left as-is (deliberate, agent-confirmed): terminalSubject's asymmetric soft-gate
(best-effort notification label — a different contract than the render consumers),
and CanvasMinimap's separate info/meta memos + inline gate (approved in the codex
pass; its geometry/badge reactivity split is load-bearing). Efficiency agent
confirmed no coarse-dep regression: every consumer reads the meta REFERENCE for
gating and leaves for content, so cadence is unchanged.

check + client unit suite green.
